+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Fix rendering of `\special{}` tags with complex contents (@klmr, #1744).
 
+* `\value{}` now does a better job of handling multiple mingled items and text 
+  (#1479).
+
 * `build_reference()` now allows linking to topics from other packages (either function names e.g. `rlang::is_installed` or topic names e.g. `sass::font_face`). (#1664)
 
 * pkgdown, for Bootstrap 4, supports tabsets in articles [as in R Markdown](https://bookdown.org/yihui/rmarkdown-cookbook/html-tabs.html) including [fading effect](https://bookdown.org/yihui/rmarkdown/html-document.html#tabbed-sections) (@JamesHWade, #1667).